+++ fractol/srcs/frames/frames_bonus.c
@@ -6,7 +6,7 @@
 /*   By: tpetit <tpetit@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/19 15:49:54 by tpetit            #+#    #+#             */
-/*   Updated: 2021/05/19 18:22:30 by tpetit           ###   ########.fr       */
+/*   Updated: 2021/05/19 18:26:04 by tpetit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,8 +36,8 @@ void	proceed_moves(t_fract *fract)
 
 int	next_frame(t_fract *fract)
 {
-	static double	last_left;
-	static double	last_top;
+	static long double	last_left;
+	static long double	last_top;
 	int				i;
 	pthread_t		t_id[THREAD_COUNT];
 	t_thread		threads[THREAD_COUNT];
